fix: move Direction enum to base/common to resolve layer violation
- Create src/vs/base/common/direction.ts with Direction enum
- Re-export Direction from base/browser/ui/grid/grid.ts for backward compatibility
- Update editorGroupsService.ts to import from common layer
- Fix resizePaneActions.ts: convert tabs to spaces, add trailing commas,
  update Direction import to common layer

The common/ layer was importing from browser/ layer (grid.ts), causing
the valid-layers-check to pull in DOM types and fail with hundreds of
errors. Moving Direction to common/ resolves the layer boundary violation.
