+++ src/vs/base/common/direction.ts
@@ -1,0 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+/**
+ * Cardinal directions used for navigating and resizing views within a grid.
+ * Placed in the common layer so it can be imported by both browser and common code.
+ */
+export const enum Direction {
+	Up,
+	Down,
+	Left,
+	Right
+}
